refactored overwrite visit command and moved removeGlobalCssTransitions
